Change TooltipTextField to a hook
