+++ portal/src/graphql/portal/SMTPConfigurationScreen.tsx
@@ -17,7 +17,7 @@ import {
   IButtonProps,
 } from "@fluentui/react";
 import { FormattedMessage, Context } from "@oursky/react-messageformat";
-import TooltipTextField from "../../TooltipTextField";
+import { useTextFieldTooltip } from "../../useTextFieldTooltip";
 import ShowError from "../../ShowError";
 import ShowLoading from "../../ShowLoading";
 import FormContainer from "../../FormContainer";
@@ -399,6 +399,20 @@ const SMTPConfigurationScreenContent: React.FC<SMTPConfigurationScreenContentPro
       [setState]
     );
 
+    const hostProps = useTextFieldTooltip({
+      tooltipLabel: renderToString("SMTPConfigurationScreen.host.tooltip"),
+    });
+
+    const portProps = useTextFieldTooltip({
+      tooltipLabel: renderToString("SMTPConfigurationScreen.port.tooltip"),
+    });
+
+    const sendgridAPIKeyProps = useTextFieldTooltip({
+      tooltipLabel: renderToString(
+        "SMTPConfigurationScreen.sendgrid.api-key.tooltip"
+      ),
+    });
+
     return (
       <ScreenContent className={styles.root}>
         <ScreenTitle>
@@ -442,7 +456,7 @@ const SMTPConfigurationScreenContent: React.FC<SMTPConfigurationScreenContentPro
                   <ProviderDescription>
                     <FormattedMessage id="SMTPConfigurationScreen.custom.description" />
                   </ProviderDescription>
-                  <TooltipTextField
+                  <TextField
                     className={styles.control}
                     type="text"
                     label={renderToString("SMTPConfigurationScreen.host.label")}
@@ -450,11 +464,9 @@ const SMTPConfigurationScreenContent: React.FC<SMTPConfigurationScreenContentPro
                     disabled={state.isPasswordMasked}
                     required={true}
                     onChange={onChangeHost}
-                    tooltipLabel={renderToString(
-                      "SMTPConfigurationScreen.host.tooltip"
-                    )}
+                    {...hostProps}
                   />
-                  <TooltipTextField
+                  <TextField
                     className={styles.control}
                     type="text"
                     label={renderToString("SMTPConfigurationScreen.port.label")}
@@ -462,9 +474,7 @@ const SMTPConfigurationScreenContent: React.FC<SMTPConfigurationScreenContentPro
                     disabled={state.isPasswordMasked}
                     required={true}
                     onChange={onChangePort}
-                    tooltipLabel={renderToString(
-                      "SMTPConfigurationScreen.port.tooltip"
-                    )}
+                    {...portProps}
                   />
                   <TextField
                     className={styles.control}
@@ -497,7 +507,7 @@ const SMTPConfigurationScreenContent: React.FC<SMTPConfigurationScreenContentPro
                   <ProviderDescription>
                     <FormattedMessage id="SMTPConfigurationScreen.sendgrid.description" />
                   </ProviderDescription>
-                  <TooltipTextField
+                  <TextField
                     className={styles.control}
                     type="password"
                     label={renderToString(
@@ -511,9 +521,7 @@ const SMTPConfigurationScreenContent: React.FC<SMTPConfigurationScreenContentPro
                     required={true}
                     disabled={state.isPasswordMasked}
                     onChange={onChangePassword}
-                    tooltipLabel={renderToString(
-                      "SMTPConfigurationScreen.sendgrid.api-key.tooltip"
-                    )}
+                    {...sendgridAPIKeyProps}
                   />
                 </>
               )}
